Rename pkg/server to pkg/sqlflowserver
