+++ cmd/sqlflowserver/e2e_hive_test.go
@@ -18,7 +18,7 @@ import (
 	"os"
 	"testing"
 
-	"sqlflow.org/sqlflow/pkg/server"
+	server "sqlflow.org/sqlflow/pkg/sqlflowserver"
 )
 
 func TestEnd2EndHive(t *testing.T) {
