+++ cmd/sqlflowserver/e2e_pai_maxcompute_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"sqlflow.org/sqlflow/pkg/server"
+	server "sqlflow.org/sqlflow/pkg/sqlflowserver"
 )
 
 func CasePAIMaxComputeTrainPredictCategoricalFeature(t *testing.T) {
